commented out original tests and added export functions to automate testing
I also couldn't split asyncCounter into 2 functions like i wanted, for some reason it was count would always be 3 no matter what the key was. While trying to fix that i saw that for some reason it also was skipping the first number in the array? unless that gets checked right as work is called?